Closes #3742:  move numeric module to arkouda.numpy
