+++ arkouda/__init__.py
@@ -13,7 +13,6 @@ from arkouda.sorting import *
 from arkouda.pdarraysetops import *
 from arkouda.pdarraycreation import *
 from arkouda.pdarraymanipulation import *
-from arkouda.numeric import *
 from arkouda.groupbyclass import *
 from arkouda.strings import *
 from arkouda.join import *
