+++ arkouda/categorical.py
@@ -23,8 +23,8 @@ from arkouda.client import generic_msg
 from arkouda.groupbyclass import GroupBy, unique
 from arkouda.infoclass import information
 from arkouda.logger import getArkoudaLogger
-from arkouda.numeric import cast as akcast
-from arkouda.numeric import where
+from arkouda.numpy import cast as akcast
+from arkouda.numpy import where
 from arkouda.numpy.dtypes import bool_ as akbool
 from arkouda.numpy.dtypes import dtype as akdtype
 from arkouda.numpy.dtypes import int64 as akint64
