+++ arkouda/client_dtypes.py
@@ -5,11 +5,11 @@ from typing import Optional, Union
 import numpy as np
 from typeguard import typechecked
 
+from arkouda.groupbyclass import GroupBy, broadcast
+from arkouda.numpy import cast as akcast
+from arkouda.numpy import where
 from arkouda.numpy.dtypes import bitType, intTypes, isSupportedInt
 from arkouda.numpy.dtypes import uint64 as akuint64
-from arkouda.groupbyclass import GroupBy, broadcast
-from arkouda.numeric import cast as akcast
-from arkouda.numeric import where
 from arkouda.pdarrayclass import RegistrationError, pdarray
 from arkouda.pdarraycreation import arange, array, create_pdarray, zeros
 from arkouda.strings import Strings
