+++ arkouda/join.py
@@ -6,11 +6,11 @@ from typeguard import typechecked
 from arkouda.alignment import right_align
 from arkouda.categorical import Categorical
 from arkouda.client import generic_msg
+from arkouda.groupbyclass import GroupBy, broadcast
+from arkouda.numpy import cumsum
 from arkouda.numpy.dtypes import NUMBER_FORMAT_STRINGS
 from arkouda.numpy.dtypes import int64 as akint64
 from arkouda.numpy.dtypes import resolve_scalar_dtype
-from arkouda.groupbyclass import GroupBy, broadcast
-from arkouda.numeric import cumsum
 from arkouda.pdarrayclass import create_pdarray, pdarray
 from arkouda.pdarraycreation import arange, array, ones, zeros
 from arkouda.pdarraysetops import concatenate, in1d
