+++ arkouda/numpy/__init__.py
@@ -95,4 +95,4 @@ from arkouda.numpy.ma import *
 from arkouda.numpy.polynomial import *
 from arkouda.numpy.rec import *
 
-from ._numeric import floor
+from ._numeric import *
