+++ arkouda/scipy/special/_math.py
@@ -3,7 +3,7 @@ from warnings import warn
 
 import numpy as np
 
-from arkouda.numeric import log
+from arkouda.numpy import log
 from arkouda.pdarrayclass import pdarray
