+++ pytest.ini
@@ -30,9 +30,8 @@ testpaths =
     tests/join_test.py
     tests/logger_test.py
     tests/message_test.py
-    tests/numeric_test.py
     tests/numpy/dtypes_test.py
-    tests/numpy/numpy_numeric_test.py
+    tests/numpy/numeric_test.py
     tests/numpy/numpy_test.py
     tests/operator_test.py
     tests/pdarray_creation_test.py
